+++ packages/cli/src/utils/transformers/transform-css-vars.ts
@@ -1,8 +1,7 @@
-import { SyntaxKind } from 'ts-morph';
-import * as z from 'zod';
-
-import { registryBaseColorSchema } from '../registry/schema';
-import { Transformer } from '../transformers';
+import { registryBaseColorSchema } from "@/src/registry/schema"
+import { Transformer } from "@/src/utils/transformers"
+import { ScriptKind, SyntaxKind } from "ts-morph"
+import { z } from "zod"
 
 export const transformCssVars: Transformer = async ({
   sourceFile,
@@ -11,12 +10,11 @@ export const transformCssVars: Transformer = async ({
 }) => {
   // No transform if using css variables.
   if (config.tailwind?.cssVariables || !baseColor?.inlineColors) {
-    return sourceFile;
+    return sourceFile
   }
 
   // Find jsx attributes with the name className.
   // const openingElements = sourceFile.getDescendantsOfKind(SyntaxKind.JsxElement)
-  // console.log(openingElements)
   // const jsxAttributes = sourceFile
   //   .getDescendantsOfKind(SyntaxKind.JsxAttribute)
   //   .filter((node) => node.getName() === "className")
@@ -32,18 +30,18 @@ export const transformCssVars: Transformer = async ({
   //   }
   // }
   sourceFile.getDescendantsOfKind(SyntaxKind.StringLiteral).forEach((node) => {
-    const value = node.getText();
+    const value = node.getText()
     if (value) {
       const valueWithColorMapping = applyColorMapping(
-        value.replace(/'/g, '').replace(/"/g, ''),
+        value.replace(/"/g, ""),
         baseColor.inlineColors
-      );
-      node.replaceWithText(`'${valueWithColorMapping.trim()}'`);
+      )
+      node.replaceWithText(`"${valueWithColorMapping.trim()}"`)
     }
-  });
+  })
 
-  return sourceFile;
-};
+  return sourceFile
+}
 
 // export default function transformer(file: FileInfo, api: API) {
 //   const j = api.jscodeshift.withParser("tsx")
@@ -61,7 +59,6 @@ export const transformCssVars: Transformer = async ({
 //       if (node?.value?.type) {
 //         if (node.value.type === "StringLiteral") {
 //           node.value.value = applyColorMapping(node.value.value)
-//           console.log(node.value.value)
 //         }
 
 //         if (
@@ -106,79 +103,79 @@ export const transformCssVars: Transformer = async ({
 // Splits a className into variant-name-alpha.
 // eg. hover:bg-primary-100 -> [hover, bg-primary, 100]
 export function splitClassName(className: string): (string | null)[] {
-  if (!className.includes('/') && !className.includes(':')) {
-    return [null, className, null];
+  if (!className.includes("/") && !className.includes(":")) {
+    return [null, className, null]
   }
 
-  const parts: (string | null)[] = [];
+  const parts: (string | null)[] = []
   // First we split to find the alpha.
-  const [rest, alpha] = className.split('/');
+  let [rest, alpha] = className.split("/")
 
   // Check if rest has a colon.
-  if (!rest.includes(':')) {
-    return [null, rest, alpha];
+  if (!rest.includes(":")) {
+    return [null, rest, alpha]
   }
 
   // Next we split the rest by the colon.
-  const split = rest.split(':');
+  const split = rest.split(":")
 
   // We take the last item from the split as the name.
-  const name = split.pop();
+  const name = split.pop()
 
   // We glue back the rest of the split.
-  const variant = split.join(':');
+  const variant = split.join(":")
 
   // Finally we push the variant, name and alpha.
-  parts.push(variant ?? null, name ?? null, alpha ?? null);
+  parts.push(variant ?? null, name ?? null, alpha ?? null)
 
-  return parts;
+  return parts
 }
 
-const PREFIXES = ['bg-', 'text-', 'border-', 'ring-offset-', 'ring-'];
+const PREFIXES = ["bg-", "text-", "border-", "ring-offset-", "ring-"]
 
 export function applyColorMapping(
   input: string,
-  mapping: z.infer<typeof registryBaseColorSchema>['inlineColors']
+  mapping: z.infer<typeof registryBaseColorSchema>["inlineColors"]
 ) {
   // Handle border classes.
-  if (input.includes(' border ')) {
-    input = input.replace(' border ', ' border border-border ');
+  if (input.includes(" border ")) {
+    input = input.replace(" border ", " border border-border ")
   }
 
   // Build color mappings.
-  const classNames = input.split(' ');
-  const lightMode: string[] = [];
-  const darkMode: string[] = [];
-  for (const className of classNames) {
-    const [variant, value, modifier] = splitClassName(className);
-    const prefix = PREFIXES.find((pre) => value?.startsWith(pre));
+  const classNames = input.split(" ")
+  const lightMode = new Set<string>()
+  const darkMode = new Set<string>()
+  for (let className of classNames) {
+    const [variant, value, modifier] = splitClassName(className)
+    const prefix = PREFIXES.find((prefix) => value?.startsWith(prefix))
     if (!prefix) {
-      if (!lightMode.includes(className)) {
-        lightMode.push(className);
+      if (!lightMode.has(className)) {
+        lightMode.add(className)
       }
-      continue;
+      continue
     }
 
-    const needle = value?.replace(prefix, '');
+    const needle = value?.replace(prefix, "")
     if (needle && needle in mapping.light) {
-      lightMode.push(
+      lightMode.add(
         [variant, `${prefix}${mapping.light[needle]}`]
           .filter(Boolean)
-          .join(':') + (modifier ? `/${modifier}` : '')
-      );
+          .join(":") + (modifier ? `/${modifier}` : "")
+      )
 
-      darkMode.push(
-        ['dark', variant, `${prefix}${mapping.dark[needle]}`]
+      darkMode.add(
+        ["dark", variant, `${prefix}${mapping.dark[needle]}`]
           .filter(Boolean)
-          .join(':') + (modifier ? `/${modifier}` : '')
-      );
-      continue;
+          .join(":") + (modifier ? `/${modifier}` : "")
+      )
+      continue
     }
 
-    if (!lightMode.includes(className)) {
-      lightMode.push(className);
+    if (!lightMode.has(className)) {
+      lightMode.add(className)
     }
   }
 
-  return lightMode.join(' ') + ' ' + darkMode.join(' ').trim();
+  return [...Array.from(lightMode), ...Array.from(darkMode)].join(" ").trim()
 }