+++ packages/dnd/src/components/Scroller/DndScroller.tsx
@@ -1,19 +1,24 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 
-import { dndStore } from '../../dndStore';
-import { Scroller, ScrollerProps } from './Scroller';
+import { usePluginOption } from '@udecode/plate/react';
+
+import { DndPlugin } from '../../DndPlugin';
+import { type ScrollerProps, Scroller } from './Scroller';
 
 export function DndScroller(props: Partial<ScrollerProps>) {
-  const isDragging = dndStore.use.isDragging();
+  const isDragging = usePluginOption(DndPlugin, 'isDragging');
+
   const [show, setShow] = React.useState(false);
 
-  useEffect(() => {
+  React.useEffect(() => {
     if (isDragging) {
       const timeout = setTimeout(() => {
         setShow(true);
       }, 100);
+
       return () => clearTimeout(timeout);
     }
+
     setShow(false);
   }, [isDragging, show]);