+++ packages/dnd/src/hooks/useDropNode.ts
@@ -1,101 +1,152 @@
-import { TEditor, TReactEditor, Value } from '@udecode/plate-common';
-import { DropTargetHookSpec, DropTargetMonitor, useDrop } from 'react-dnd';
+import {
+  type DropTargetHookSpec,
+  type DropTargetMonitor,
+  useDrop,
+} from 'react-dnd';
 
-import { onDropNode } from '../transforms/onDropNode';
+import type { NodeEntry, TElement } from '@udecode/plate';
+import type { PlateEditor } from '@udecode/plate/react';
+
+import type {
+  DragItemNode,
+  ElementDragItemNode,
+  FileDragItemNode,
+} from '../types';
+
+import { DndPlugin } from '../DndPlugin';
+import { getDropPath, onDropNode } from '../transforms/onDropNode';
 import { onHoverNode } from '../transforms/onHoverNode';
-import { DragItemNode, DropLineDirection } from '../types';
+
+export type CanDropCallback = (args: {
+  dragEntry: NodeEntry<TElement>;
+  dragItem: DragItemNode;
+  dropEntry: NodeEntry<TElement>;
+  editor: PlateEditor;
+}) => boolean;
 
 export interface UseDropNodeOptions
   extends DropTargetHookSpec<DragItemNode, unknown, { isOver: boolean }> {
-  /**
-   * The reference to the node being dragged.
-   */
+  /** The node to which the drop line is attached. */
+  element: TElement;
+
+  /** The reference to the node being dragged. */
   nodeRef: any;
 
   /**
-   * Id of the node.
+   * Intercepts the drop handling. If `false` is returned, the default drop
+   * behavior is called after. If `true` is returned, the default behavior is
+   * not called.
    */
-  id: string;
+  canDropNode?: CanDropCallback;
 
-  /**
-   * Current value of dropLine.
-   */
-  dropLine: string;
+  orientation?: 'horizontal' | 'vertical';
 
-  /**
-   * Callback called on dropLine change.
-   */
-  onChangeDropLine: (newValue: DropLineDirection) => void;
-  /**
-   * Intercepts the drop handling.
-   * If `false` is returned, the default drop behavior is called after.
-   * If `true` is returned, the default behavior is not called.
-   */
   onDropHandler?: (
-    editor: TEditor,
+    editor: PlateEditor,
     props: {
-      monitor: DropTargetMonitor<DragItemNode, unknown>;
-      dragItem: DragItemNode;
-      nodeRef: any;
       id: string;
+      dragItem: DragItemNode;
+      monitor: DropTargetMonitor<DragItemNode, unknown>;
+      nodeRef: any;
     }
-  ) => boolean;
+  ) => boolean | void;
 }
 
 /**
  * `useDrop` hook to drop a node on the editor.
  *
  * On drop:
- * - get hover direction (top, bottom or undefined), return early if undefined
- * - dragPath: find node with id = dragItem.id, return early if not found
- * - focus editor
- * - dropPath: find node with id = id, its path should be next (bottom) or previous (top)
- * - move node from dragPath to dropPath
+ *
+ * - Get hover direction (top, bottom or undefined), return early if undefined
+ * - DragPath: find node with id = dragItem.id, return early if not found
+ * - Focus editor
+ * - DropPath: find node with id = id, its path should be next (bottom) or
+ *   previous (top)
+ * - Move node from dragPath to dropPath
  *
  * On hover:
- * - get drop line direction
- * - if differs from dropLine, setDropLine is called
+ *
+ * - Get drop line direction
+ * - If differs from dropLine, setDropLine is called
  *
  * Collect:
- * - isOver: true if mouse is over the block
+ *
+ * - IsOver: true if mouse is over the block
  */
-export const useDropNode = <V extends Value>(
-  editor: TReactEditor<V>,
+export const useDropNode = (
+  editor: PlateEditor,
   {
+    canDropNode,
+    element,
     nodeRef,
-    id,
-    dropLine,
-    onChangeDropLine,
+    orientation,
     onDropHandler,
     ...options
   }: UseDropNodeOptions
 ) => {
+  const id = element.id as string;
+
   return useDrop<DragItemNode, unknown, { isOver: boolean }>({
+    collect: (monitor) => ({
+      isOver: monitor.isOver({
+        shallow: true,
+      }),
+    }),
     drop: (dragItem, monitor) => {
+      // Don't call onDropNode if this is a file drop
+
+      if (!(dragItem as ElementDragItemNode).id) {
+        const result = getDropPath(editor, {
+          canDropNode,
+          dragItem,
+          element,
+          monitor,
+          nodeRef,
+          orientation,
+        });
+
+        const onDropFiles = editor.getOptions(DndPlugin).onDropFiles;
+
+        if (!result || !onDropFiles) return;
+
+        return onDropFiles({
+          id,
+          dragItem: dragItem as FileDragItemNode,
+          editor,
+          monitor,
+          nodeRef,
+          target: result.to,
+        });
+      }
+
       const handled =
         !!onDropHandler &&
         onDropHandler(editor, {
-          nodeRef,
           id,
           dragItem,
           monitor,
+          nodeRef,
         });
 
       if (handled) return;
 
-      onDropNode(editor, { nodeRef, id, dragItem, monitor });
+      onDropNode(editor, {
+        canDropNode,
+        dragItem: dragItem as ElementDragItemNode,
+        element,
+        monitor,
+        nodeRef,
+        orientation,
+      });
     },
-    collect: (monitor) => ({
-      isOver: monitor.isOver(),
-    }),
     hover(item: DragItemNode, monitor: DropTargetMonitor) {
       onHoverNode(editor, {
-        nodeRef,
-        id,
-        dropLine,
-        onChangeDropLine,
+        canDropNode,
         dragItem: item,
+        element,
         monitor,
+        nodeRef,
+        orientation,
       });
     },
     ...options,