+++ packages/mention/src/handlers/KeyboardEventHandler.ts
@@ -1,5 +1,0 @@
-import { KeyboardEvent } from 'react';
-import { HandlerReturnType } from '@udecode/plate-common';
-
-// TODO: move to core
-export type KeyboardEventHandler = (event: KeyboardEvent) => HandlerReturnType;