+++ packages/slate-utils/src/queries/isSelectionExpanded.ts
@@ -1,7 +1,0 @@
-import { isExpanded, TEditor, Value } from '@udecode/slate';
-
-/**
- * Is the selection expanded.
- */
-export const isSelectionExpanded = <V extends Value>(editor: TEditor<V>) =>
-  isExpanded(editor.selection);