+++ packages/test-utils/src/jsx.ts
@@ -1,10 +1,14 @@
-/* eslint-disable prettier/prettier */
-// @ts-ignore
-import { createHyperscript, createText as createTestText } from "slate-hyperscript";
-import { HyperscriptShorthands } from "slate-hyperscript/dist/hyperscript";
-import { createText } from "./hyperscript/creators";
+import {
+  type HyperscriptShorthands,
+  createHyperscript as createHyperscriptBase,
+  createText as createTestText,
+} from 'slate-hyperscript';
+
+import { createText } from './internals/creators';
+import { createHyperscript } from './internals/hyperscript';
 
 declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace JSX {
     interface IntrinsicElements {
       [key: string]: any;
@@ -15,91 +19,85 @@ declare global {
       element: any;
       focus: any;
       fragment: any;
-      selection: any;
+      hli: any;
+      hol: any;
+      hp: any;
       htext: {
         [key: string]: any;
         // These optional params will show up in the autocomplete!
         bold?: boolean;
-        underline?: boolean;
-        italic?: boolean;
-        code?: boolean;
         children?: any;
+        code?: boolean;
+        italic?: boolean;
+        underline?: boolean;
       };
-      hp: any;
       hul: any;
-      hol: any;
-      hli: any;
+      selection: any;
     }
   }
 }
 
 const voidChildren = [{ text: '' }];
 
-const ELEMENT_BLOCKQUOTE = 'blockquote';
-const ELEMENT_CODE_BLOCK = 'code_block';
-const ELEMENT_CODE_LINE = 'code_line';
-const ELEMENT_EXCALIDRAW = 'excalidraw';
-const ELEMENT_H1 = 'h1';
-const ELEMENT_H2 = 'h2';
-const ELEMENT_H3 = 'h3';
-const ELEMENT_H4 = 'h4';
-const ELEMENT_H5 = 'h5';
-const ELEMENT_H6 = 'h6';
-const ELEMENT_IMAGE = 'img';
-const ELEMENT_LI = 'li';
-const ELEMENT_LIC = 'lic';
-const ELEMENT_LINK = 'a';
-const ELEMENT_MEDIA_EMBED = 'media_embed';
-const ELEMENT_MENTION = 'mention';
-const ELEMENT_MENTION_INPUT = 'mention_input';
-const ELEMENT_OL = 'ol';
-const ELEMENT_PARAGRAPH = 'p';
-const ELEMENT_TABLE = 'table';
-const ELEMENT_TD = 'td';
-const ELEMENT_TH = 'th';
-const ELEMENT_TODO_LI = 'action_item';
-const ELEMENT_TR = 'tr';
-const ELEMENT_UL = 'ul';
-
 const elements: HyperscriptShorthands = {
-  ha: { type: ELEMENT_LINK },
-  hblockquote: { type: ELEMENT_BLOCKQUOTE },
-  hcodeblock: { type: ELEMENT_CODE_BLOCK },
-  hcodeline: { type: ELEMENT_CODE_LINE },
-  hdefault: { type: ELEMENT_PARAGRAPH },
-  hexcalidraw: { type: ELEMENT_EXCALIDRAW },
-  hh1: { type: ELEMENT_H1 },
-  hh2: { type: ELEMENT_H2 },
-  hh3: { type: ELEMENT_H3 },
-  hh4: { type: ELEMENT_H4 },
-  hh5: { type: ELEMENT_H5 },
-  hh6: { type: ELEMENT_H6 },
-  himg: { type: ELEMENT_IMAGE, children: voidChildren },
-  hli: { type: ELEMENT_LI },
-  hlic: { type: ELEMENT_LIC },
-  hmediaembed: { type: ELEMENT_MEDIA_EMBED, children: voidChildren },
-  hmention: { type: ELEMENT_MENTION, children: voidChildren },
-  hmentioninput: { type: ELEMENT_MENTION_INPUT, children: voidChildren },
-  hol: { type: ELEMENT_OL },
-  hp: { type: ELEMENT_PARAGRAPH },
-  htable: { type: ELEMENT_TABLE },
-  htd: { type: ELEMENT_TD },
-  hth: { type: ELEMENT_TH },
-  htodoli: { type: ELEMENT_TODO_LI },
-  htr: { type: ELEMENT_TR },
-  hul: { type: ELEMENT_UL },
+  ha: { type: 'a' },
+  haudio: { children: voidChildren, type: 'audio' },
+  hblockquote: { type: 'blockquote' },
+  hcallout: { type: 'callout' },
+  hcodeblock: { type: 'code_block' },
+  hcodeline: { type: 'code_line' },
+  hcolumn: { type: 'column' },
+  hcolumngroup: { type: 'column_group' },
+  hdate: { children: voidChildren, type: 'date' },
+  hdefault: { type: 'p' },
+  hequation: { type: 'equation' },
+  hexcalidraw: { type: 'excalidraw' },
+  hfile: { children: voidChildren, type: 'file' },
+  hh1: { type: 'h1' },
+  hh2: { type: 'h2' },
+  hh3: { type: 'h3' },
+  hh4: { type: 'h4' },
+  hh5: { type: 'h5' },
+  hh6: { type: 'h6' },
+  himg: { children: voidChildren, type: 'img' },
+  hinlineequation: { type: 'inline_equation' },
+  hli: { type: 'li' },
+  hlic: { type: 'lic' },
+  hmediaembed: { children: voidChildren, type: 'media_embed' },
+  hmention: { children: voidChildren, type: 'mention' },
+  hmentioninput: { children: voidChildren, type: 'mention_input' },
+  hnli: { type: 'nli' },
+  hol: { type: 'ol' },
+  hp: { type: 'p' },
+  hplaceholder: { children: voidChildren, type: 'placeholder' },
+  htable: { type: 'table' },
+  htd: { type: 'td' },
+  hth: { type: 'th' },
+  htoc: { type: 'toc' },
+  htodoli: { type: 'action_item' },
+  htoggle: { type: 'toggle' },
+  htr: { type: 'tr' },
+  hul: { type: 'ul' },
+  hvideo: { children: voidChildren, type: 'video' },
 };
 
 export const jsx = createHyperscript({
-  elements,
   creators: {
     htext: createTestText,
   },
+  elements,
+});
+
+export const jsxt = createHyperscriptBase({
+  creators: {
+    htext: createTestText,
+  },
+  elements,
 });
 
 export const hjsx = createHyperscript({
-  elements,
   creators: {
     htext: createText,
   },
+  elements,
 });