+++ packages/utils/src/types/index.ts
@@ -3,7 +3,8 @@
  */
 
 export * from './AnyObject';
-export * from './DeepPartial';
+export * from './Deep';
 export * from './FunctionProperties';
+export * from './Nullable';
 export * from './WithOptional';
 export * from './types';